Textures WIP. Implemented validation layers. Architectural changes.

+++ Triangle/include/VulkanRenderPassCreateInfo.h
@@ -14,7 +14,7 @@ public:
 
 	VulkanRenderPassCreateInfo& setAttachments(std::vector<VkAttachmentDescription>& attachments)
 	{
-		vkInfo.attachmentCount = attachments.size();
+		vkInfo.attachmentCount = static_cast<uint32_t>(attachments.size());
 		vkInfo.pAttachments = attachments.data();
 		return *this;
 	}
